+++ src/views/partials/GitCredentialDetailsModal.vue
